# openapi.yaml
openapi: "3.0.3"
info:
  title: Printing Press Manager API
  version: "1.0.0"
  description: |
    Backend server for a kiosk-based printing press system.
    Manages kiosks, documents, jobs, payments, and print commands.
servers:
  - url: http://localhost:8080
    description: Local development
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Payment:
      type: object
      properties:
        id:
          type: string
        jobId:
          type: string
        amount:
          type: number
        currency:
          type: string
        provider:
          type: string
        status:
          type: string
          enum: [pending, paid, failed]
        paymentLink:
          type: string
        createdAt:
          type: string
          format: date-time
        paidAt:
          type: string
          format: date-time
          nullable: true
paths:
...
  /payments:
    post:
      summary: Generate a payment link for a job
      tags: [Payments]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [jobId]
              properties:
                jobId:
                  type: string
                provider:
                  type: string
                  description: Payment provider (razorpay, stripe, etc.)
      responses:
        "201":
          description: Payment link created
          content:
            application/json:
              schema:
                type: object
                properties:
                  paymentId:
                    type: string
                  paymentLink:
                    type: string
                  amount:
                    type: number
                  currency:
                    type: string
  /payments/{paymentId}:
    get:
      summary: Get payment status
      tags: [Payments]
      parameters:
        - name: paymentId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Payment details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
  /payments/{paymentId}/webhook:
    post:
      summary: Payment gateway webhook callback
      tags: [Payments]
      parameters:
        - name: paymentId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  type: string
                  enum: [success, failed]
      responses:
        "200":
          description: Webhook processed
          content:
            application/json:
              schema:
                type: object
                properties:
                  received:
                    type: boolean
                  paymentStatus:
                    type: string